rename erc20-contracts to erc20-metadata-functions
