+++ Cargo.toml
@@ -2,12 +2,11 @@
 members = [
   "erc20",
   "erc20-metadata",
+  "erc20-metadata-functions",
   "erc20-stores",
   "erc20-balances",
   "erc20-supply",
   "erc20-balances-rpc",
-  "erc20-contracts",
-  "erc20-contracts-rpc",
   "native-balances",
   "native-contracts",
   "uniswap-v2",
