--- erc20-contracts-rpc/README.md
@@ -1,3 +1,0 @@
-# Substreams: `ERC-20 Contracts` via RPC
-
-> Extracts ERC-20 token contract metadata from the Ethereum blockchain.
