--- erc20-contracts-rpc/src/lib.rs
@@ -1,3 +1,0 @@
-pub mod calls;
-pub mod maps;
-pub mod stores;
