--- erc20-contracts-rpc/substreams.yaml
@@ -1,39 +1,0 @@
-specVersion: v0.1.0
-package:
-  name: evm_erc20_contracts_rpc
-  version: v1.11.0
-  url: https://githubcom/pinax-network/substreams-evm-tokens
-  description: EVM Token contracts via RPC
-  image: ../image.png
-
-imports:
-  v1_10: https://github.com/pinax-network/substreams-evm-tokens/releases/download/v1.10.0/evm-tokens-erc20-contracts-rpc-v1.10.0.spkg
-
-binaries:
-  default:
-    type: wasm/rust-v1
-    file: ../target/wasm32-unknown-unknown/release/erc20_contracts_rpc.wasm
-
-protobuf:
-  files:
-    - v1/contracts.proto
-  importPaths:
-    - ../proto
-
-modules:
-  - name: map_events
-    kind: map
-    inputs:
-      - store: v1_10:store_erc20_transfers
-        mode: deltas
-    output:
-      type: proto:evm.tokens.contracts.v1.Events
-
-  - name: store_erc20_transfers
-    kind: store
-    updatePolicy: add
-    valueType: bigint
-    inputs:
-      - source: sf.ethereum.type.v2.Block
-
-network: mainnet
