--- erc20-contracts/Cargo.toml
@@ -1,29 +1,0 @@
-[package]
-name = "erc20-contracts"
-description = { workspace = true }
-edition = { workspace = true }
-version = { workspace = true }
-authors = [
-    "Denis <denis@pinax.network>",
-    "Etienne Donneger <etienne.d@pinax.network>",
-    "MathieuLefebvre <mathieu@pinax.network>"
-]
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-
-[dependencies]
-substreams = { workspace = true }
-substreams-ethereum = { workspace = true }
-substreams-abis = { workspace = true }
-proto = { path = "../proto" }
-common = { path = "../common" }
-ethabi = "18.0"
-prost = { workspace = true }
-prost-types = { workspace = true }
-num-bigint = { version = "0.4", features = ["serde"] }
-
-[build-dependencies]
-prost-build = "0.13"
-anyhow = "1"
-substreams-ethereum = "0.10"
