--- erc20-contracts/Makefile
@@ -1,17 +1,0 @@
-.PHONY: all
-all:
-	make build
-
-.PHONY: build
-build:
-	cargo build --target wasm32-unknown-unknown --release
-	substreams pack
-
-.PHONY: parquet
-parquet:
-	rm -f state.yaml && substreams-sink-files run eth.substreams.pinax.network:443 ./substreams.yaml map_events "./out" 21525891: --encoder parquet --file-block-count 100
-
-.PHONY: gui
-gui: build
-	substreams gui substreams.yaml map_events -e eth.substreams.pinax.network:443 --network eth -s 21525891
-
