--- erc20-contracts/README.md
@@ -1,3 +1,0 @@
-# Substreams: `ERC-20 Contracts`
-
-> Extracts ERC-20 token contract metadata from the Ethereum blockchain.
