--- erc20-contracts/src/lib.rs
@@ -1,3 +1,0 @@
-pub mod maps;
-pub mod metadata;
-pub mod abi;
