+++ erc20-metadata-functions/src/functions.rs
@@ -1,12 +1,11 @@
 use crate::abi::functions;
-use common::{bigint_to_uint8, bytes32_to_string};
+use common::bytes32_to_string;
 use substreams_abis::evm::tokens;
 use substreams_ethereum::{pb::eth::v2::Call, Function};
 
 pub struct SetNameSymbol {
     pub name: Option<String>,
     pub symbol: Option<String>,
-    pub decimals: Option<i32>,
 }
 
 pub fn get_metadata<'a>(call: &'a Call) -> Option<SetNameSymbol> {
@@ -15,7 +14,6 @@ pub fn get_metadata<'a>(call: &'a Call) -> Option<SetNameSymbol> {
         return Some(SetNameSymbol {
             name: Some(name),
             symbol: Some(symbol),
-            decimals: None,
         });
     }
 
@@ -24,7 +22,6 @@ pub fn get_metadata<'a>(call: &'a Call) -> Option<SetNameSymbol> {
         return Some(SetNameSymbol {
             name: Some(name),
             symbol: None,
-            decimals: None,
         });
     }
 
@@ -33,7 +30,6 @@ pub fn get_metadata<'a>(call: &'a Call) -> Option<SetNameSymbol> {
         return Some(SetNameSymbol {
             name: None,
             symbol: Some(symbol),
-            decimals: None,
         });
     }
     None
@@ -106,26 +102,6 @@ pub fn get_name_symbol<'a>(call: &'a Call) -> Option<(String, String)> {
     // https://etherscan.io/tx/0xc83d0363850e3cbf82d04df615f3d67ccb1274cfb0944a7bd383ad89e5d3f67f
     if let Some(result) = tokens::blendr::functions::Launch::match_and_decode(call) {
         return Some((result.name, result.symbol));
-    }
-    None
-}
-
-pub fn get_name_symbol_precision<'a>(call: &'a Call) -> Option<(String, String, i32)> {
-    // USDC Circle
-    if let Some(result) = tokens::usdc::functions::Initialize::match_and_decode(call) {
-        if let Some(decimals) = bigint_to_uint8(&result.token_decimals) {
-            return Some((result.token_name, result.token_symbol, decimals));
-        }
-    }
-
-    // Wormhole
-    // TO-DO: More validation needed
-    // https://github.com/pinax-network/substreams-evm-tokens/issues/28
-    // https://github.com/wormhole-foundation/wormhole/blob/9b9c07b4b1474a1526c9a7346dab640e627630bc/ethereum/contracts/bridge/token/TokenImplementation.sol#L17-L25
-    if let Some(result) = functions::Initialize::match_and_decode(call) {
-        if let Some(decimals) = bigint_to_uint8(&result.decimals) {
-            return Some((result.name, result.symbol, decimals));
-        }
     }
     None
 }
