+++ erc20-metadata/substreams.yaml
@@ -3,7 +3,7 @@ package:
   name: erc20_metadata
   version: v0.1.0
   url: https://github.com/pinax-network/substreams-evm-tokens
-  description: ERC20 token metadata
+  description: ERC20 Token Metadata via RPC
   image: ../image.png
 
 imports:
